DEV: Unify code between PdfReader and PdfWriter (#2497)
