+++ docs/modules/PdfDocCommon.md
@@ -1,0 +1,5 @@
+# The PdfDocCommon Class
+
+**PdfDocCommon** is an abstract class which is inherited by **PdfReader** and **PdfWriter**
+
+Where indentified in the API, you can use any of the derivated class.
