--- docs/modules/PdfMerger.rst
@@ -1,7 +1,0 @@
-The PdfMerger Class
--------------------
-
-.. autoclass:: pypdf.PdfMerger
-    :members:
-    :undoc-members:
-    :show-inheritance:
