+++ docs/modules/PdfReader.rst
@@ -3,6 +3,7 @@ The PdfReader Class
 
 .. autoclass:: pypdf.PdfReader
     :members:
+    :inherited-members:
     :undoc-members:
     :show-inheritance:
