+++ docs/modules/generic.rst
@@ -7,10 +7,23 @@ Generic PDF objects
     :show-inheritance:
     :exclude-members: Destination, Field, Fit, RectangleObject
 
+
 .. autoclass:: pypdf._protocols.PdfObjectProtocol
     :members:
     :undoc-members:
     :show-inheritance:
+
+
+.. autoclass:: pypdf._protocols.XmpInformationProtocol
+    :members:
+    :undoc-members:
+    :show-inheritance:
+
+
+.. autoclass:: pypdf._protocols.PdfCommonDocProtocol
+       :members:
+       :undoc-members:
+       :show-inheritance:
 
 
 .. autoclass:: pypdf._protocols.PdfReaderProtocol
