+++ pypdf/__init__.py
@@ -8,10 +8,11 @@ You can read the full docs at https://pypdf.readthedocs.io/.
 """
 
 from ._crypt_providers import crypt_provider
+from ._doc_common import DocumentInformation
 from ._encryption import PasswordType
 from ._merger import PdfMerger
 from ._page import PageObject, Transformation, mult
-from ._reader import DocumentInformation, PdfReader
+from ._reader import PdfReader
 from ._version import __version__
 from ._writer import ObjectDeletionFlag, PdfWriter
 from .constants import ImageType
