+++ pypdf/_writer.py
@@ -54,8 +54,9 @@ from typing import (
 )
 
 from ._cmap import build_char_map_from_dict
+from ._doc_common import PdfDocCommon
 from ._encryption import EncryptAlgorithm, Encryption
-from ._page import PageObject, _VirtualList
+from ._page import PageObject
 from ._page_labels import nums_clear_range, nums_insert, nums_next
 from ._reader import PdfReader
 from ._utils import (
@@ -96,7 +97,6 @@ from .generic import (
     DecodedStreamObject,
     Destination,
     DictionaryObject,
-    Field,
     Fit,
     FloatObject,
     IndirectObject,
@@ -121,6 +121,7 @@ from .types import (
     OutlineType,
     PagemodeType,
 )
+from .xmp import XmpInformation
 
 OPTIONAL_READ_WRITE_FIELD = FieldFlag(0)
 ALL_DOCUMENT_PERMISSIONS = UserAccessPermissions.all()
@@ -146,7 +147,7 @@ def _rolling_checksum(stream: BytesIO, blocksize: int = 65536) -> str:
     return hash.hexdigest()
 
 
-class PdfWriter:
+class PdfWriter(PdfDocCommon):
     """
     Write a PDF file out, given pages produced by another class or through
     cloning a PDF file during initialization.
@@ -154,13 +155,23 @@ class PdfWriter:
     Typically data is added from a :class:`PdfReader<pypdf.PdfReader>`.
     """
 
+    # for commonality
+    @property
+    def is_encrypted(self) -> bool:
+        """
+        Read-only boolean property showing whether this PDF file is encrypted.
+
+        Note that this property, if true, will remain true even after the
+        :meth:`decrypt()<pypdf.PdfReader.decrypt>` method is called.
+        """
+        return False
+
     def __init__(
         self,
         fileobj: StrByteType = "",
         clone_from: Union[None, PdfReader, StrByteType, Path] = None,
     ) -> None:
         self._header = b"%PDF-1.3"
-
         self._objects: List[PdfObject] = []
         """The indirect objects in the PDF."""
 
@@ -179,6 +190,7 @@ class PdfWriter:
             }
         )
         self._pages = self._add_object(pages)
+        self.flattened_pages = []
 
         # info object
         info = DictionaryObject()
@@ -189,7 +201,7 @@ class PdfWriter:
                 )
             }
         )
-        self._info = self._add_object(info)
+        self._info_obj: PdfObject = self._add_object(info)
 
         # root object
         self._root_object = DictionaryObject()
@@ -222,6 +234,38 @@ class PdfWriter:
         """
         return self._root_object
 
+    @property
+    def _info(self) -> Optional[DictionaryObject]:
+        """
+        Provide access to "/Info". standardized with PdfWriter.
+
+        Returns:
+            /Info Dictionary ; None if the entry does not exists
+        """
+        return cast(DictionaryObject, self._info_obj.get_object())
+
+    @_info.setter
+    def _info(self, value: Union[IndirectObject, DictionaryObject]) -> None:
+        obj = cast(DictionaryObject, self._info_obj.get_object())
+        obj.clear()
+        obj.update(cast(DictionaryObject, value.get_object()))
+
+    @property
+    def xmp_metadata(self) -> Optional[XmpInformation]:
+        """XMP (Extensible Metadata Platform) data."""
+        return cast(XmpInformation, self.root_object.xmp_metadata)
+
+    @xmp_metadata.setter
+    def xmp_metadata(self, value: Optional[XmpInformation]) -> None:
+        """XMP (Extensible Metadata Platform) data."""
+        if value is None:
+            if "/Metadata" in self.root_object:
+                del self.root_object["/Metadata"]
+        else:
+            self.root_object[NameObject("/Metadata")] = value
+
+        return self.root_object.xmp_metadata  # type: ignore
+
     def __enter__(self) -> "PdfWriter":
         """Store that writer is initialized by 'with'."""
         self.with_as_usage = True
@@ -237,23 +281,60 @@ class PdfWriter:
         if self.fileobj:
             self.write(self.fileobj)
 
-    @property
-    def pdf_header(self) -> bytes:
+    def _repr_mimebundle_(
+        self,
+        include: Union[None, Iterable[str]] = None,
+        exclude: Union[None, Iterable[str]] = None,
+    ) -> Dict[str, Any]:
         """
+        Integration into Jupyter Notebooks.
+
+        This method returns a dictionary that maps a mime-type to it's
+        representation.
+
+        See https://ipython.readthedocs.io/en/stable/config/integrating.html
+        """
+        pdf_data = BytesIO()
+        self.write(pdf_data)
+        data = {
+            "application/pdf": pdf_data,
+        }
+
+        if include is not None:
+            # Filter representations based on include list
+            data = {k: v for k, v in data.items() if k in include}
+
+        if exclude is not None:
+            # Remove representations based on exclude list
+            data = {k: v for k, v in data.items() if k not in exclude}
+
+        return data
+
+    @property
+    def pdf_header(self) -> str:
+        """
+        Read/Write Property
         Header of the PDF document that is written.
 
-        This should be something like ``b'%PDF-1.5'``. It is recommended to set
+        This should be something like ``'%PDF-1.5'``. It is recommended to set
         the lowest version that supports all features which are used within the
         PDF file.
+
+        Note: `pdf_header` returns a string but accepts bytes or str for writing
         """
-        return self._header
+        return self._header.decode()
 
     @pdf_header.setter
-    def pdf_header(self, new_header: bytes) -> None:
+    def pdf_header(self, new_header: Union[str, bytes]) -> None:
+        if isinstance(new_header, str):
+            new_header = new_header.encode()
         self._header = new_header
 
     def _add_object(self, obj: PdfObject) -> IndirectObject:
-        if hasattr(obj, "indirect_reference") and obj.indirect_reference.pdf == self:  # type: ignore
+        if (
+            getattr(obj, "indirect_reference", None) is not None
+            and obj.indirect_reference.pdf == self  # type: ignore
+        ):
             return obj.indirect_reference  # type: ignore
         # check for /Contents in Pages (/Contents in annotation are strings)
         if isinstance(obj, DictionaryObject) and isinstance(
@@ -284,6 +365,11 @@ class PdfWriter:
                 raise ValueError("pdf must be self")
             indirect_reference = indirect_reference.idnum
         gen = self._objects[indirect_reference - 1].indirect_reference.generation  # type: ignore
+        if (
+            getattr(obj, "indirect_reference", None) is not None
+            and obj.indirect_reference.pdf != self  # type: ignore
+        ):
+            obj = obj.clone(self)
         self._objects[indirect_reference - 1] = obj
         obj.indirect_reference = IndirectObject(indirect_reference, gen, self)
         return self._objects[indirect_reference - 1]
@@ -291,7 +377,7 @@ class PdfWriter:
     def _add_page(
         self,
         page: PageObject,
-        action: Callable[[Any, IndirectObject], None],
+        action: Callable[[Any, Union[PageObject, IndirectObject]], None],
         excluded_keys: Iterable[str] = (),
     ) -> PageObject:
         assert cast(str, page[PA.TYPE]) == CO.PAGE
@@ -311,13 +397,12 @@ class PdfWriter:
         page = cast("PageObject", page_org.clone(self, False, excluded_keys))
         if page_org.pdf is not None:
             other = page_org.pdf.pdf_header
-            if isinstance(other, str):
-                other = other.encode()
             self.pdf_header = _get_max_pdf_version_header(self.pdf_header, other)
         page[NameObject(PA.PARENT)] = self._pages
         pages = cast(DictionaryObject, self.get_object(self._pages))
         assert page.indirect_reference is not None
         action(pages[PA.KIDS], page.indirect_reference)
+        action(self.flattened_pages, page)
         page_count = cast(int, pages[PA.COUNT])
         pages[NameObject(PA.COUNT)] = NumberObject(page_count + 1)
         return page
@@ -353,21 +438,6 @@ class PdfWriter:
             logger_warning(
                 f"set_need_appearances_writer({state}) catch : {exc}", __name__
             )
-
-    @property
-    def viewer_preferences(self) -> Optional[ViewerPreferences]:
-        """Returns the existing ViewerPreferences as an overloaded dictionary."""
-        o = self._root_object.get(CD.VIEWER_PREFERENCES, None)
-        if o is None:
-            return None
-        o = o.get_object()
-        if not isinstance(o, ViewerPreferences):
-            o = ViewerPreferences(o)
-            if hasattr(o, "indirect_reference"):
-                self._replace_object(o.indirect_reference, o)
-            else:
-                self._root_object[NameObject(CD.VIEWER_PREFERENCES)] = o
-        return o
 
     def create_viewer_preferences(self) -> ViewerPreferences:
         o = ViewerPreferences()
@@ -417,25 +487,6 @@ class PdfWriter:
         """
         return self._add_page(page, lambda kids, p: kids.insert(index, p))
 
-    def get_page(self, page_number: int) -> PageObject:
-        """
-        Retrieve a page by number from this PDF file.
-
-        Args:
-            page_number: The page number to retrieve
-                (pages begin at zero)
-
-        Returns:
-            The page at the index given by *page_number*
-        """
-        pages = cast(Dict[str, Any], self.get_object(self._pages))
-        # TODO: crude hack
-        return cast(PageObject, pages[PA.KIDS][page_number].get_object())
-
-    def _get_num_pages(self) -> int:
-        pages = cast(Dict[str, Any], self.get_object(self._pages))
-        return int(pages[NameObject("/Count")])
-
     def _get_page_number_by_indirect(
         self, indirect_reference: Union[None, int, NullObject, IndirectObject]
     ) -> Optional[int]:
@@ -457,22 +508,6 @@ class PdfWriter:
         if isinstance(obj, PageObject):
             return obj.page_number
         return None
-
-    @property
-    def pages(self) -> List[PageObject]:
-        """
-        Property that emulates a list of :class:`PageObject<pypdf._page.PageObject>`.
-        this property allows to get a page or  a range of pages.
-
-        It also provides capability to remove a page or range of pages from
-        the list (through del operator).
-        Note: only the page entry is removed, as the objects beneath can be used
-        elsewhere.
-        A solution to completely remove them - if they are not used somewhere -
-        is to write to a buffer/temporary and to then load it into a new PdfWriter
-        object.
-        """
-        return _VirtualList(self._get_num_pages, self.get_page)  # type: ignore
 
     def add_blank_page(
         self, width: Optional[float] = None, height: Optional[float] = None
@@ -523,7 +558,7 @@ class PdfWriter:
             PageSizeNotDefinedError: if width and height are not defined
                 and previous page does not exist.
         """
-        if width is None or height is None and (self._get_num_pages() - 1) >= index:
+        if width is None or height is None and (self.get_num_pages() - 1) >= index:
             oldpage = self.pages[index]
             width = oldpage.mediabox.width
             height = oldpage.mediabox.height
@@ -535,29 +570,7 @@ class PdfWriter:
     def open_destination(
         self,
     ) -> Union[None, Destination, TextStringObject, ByteStringObject]:
-        """
-        Property to access the opening destination (``/OpenAction`` entry in
-        the PDF catalog). It returns ``None`` if the entry does not exist is not
-        set.
-
-        Raises:
-            Exception: If a destination is invalid.
-        """
-        if "/OpenAction" not in self._root_object:
-            return None
-        oa = self._root_object["/OpenAction"]
-        if isinstance(oa, (str, bytes)):
-            return create_string_object(str(oa))
-        elif isinstance(oa, ArrayObject):
-            try:
-                page, typ = oa[0:2]
-                array = oa[2:]
-                fit = Fit(typ, tuple(array))
-                return Destination("OpenAction", page, fit)
-            except Exception as exc:
-                raise Exception(f"Invalid Destination {oa}: {exc}")
-        else:
-            return None
+        return super().open_destination
 
     @open_destination.setter
     def open_destination(self, dest: Union[None, str, Destination, PageObject]) -> None:
@@ -733,19 +746,6 @@ class PdfWriter:
             # Trigger callback, pass writer page as parameter
             if callable(after_page_append):
                 after_page_append(writer_page)
-
-    def _get_qualified_field_name(self, parent: DictionaryObject) -> Optional[str]:
-        if "/TM" in parent:
-            return cast(str, parent["/TM"])
-        elif "/T" not in parent:
-            return None
-        elif "/Parent" in parent:
-            qualified_parent = self._get_qualified_field_name(
-                cast(DictionaryObject, parent["/Parent"])
-            )
-            if qualified_parent is not None:
-                return qualified_parent + "." + cast(str, parent["/T"])
-        return cast(str, parent["/T"])
 
     def _update_text_field(self, field: DictionaryObject) -> None:
         # Calculate rectangle dimensions
@@ -1016,76 +1016,6 @@ class PdfWriter:
                 lst.append(ano)
         return lst
 
-    def get_pages_showing_field(
-        self, field: Union[Field, PdfObject, IndirectObject]
-    ) -> List[PageObject]:
-        """
-        Provides list of pages where the field is called.
-
-        Args:
-            field: Field Object, PdfObject or IndirectObject referencing a Field
-
-        Returns:
-            List of pages:
-                - Empty list:
-                    The field has no widgets attached
-                    (either hidden field or ancestor field).
-                - Single page list:
-                    Page where the widget is present
-                    (most common).
-                - Multi-page list:
-                    Field with multiple kids widgets
-                    (example: radio buttons, field repeated on multiple pages).
-        """
-
-        def _get_inherited(obj: DictionaryObject, key: str) -> Any:
-            if key in obj:
-                return obj[key]
-            elif "/Parent" in obj:
-                return _get_inherited(
-                    cast(DictionaryObject, obj["/Parent"].get_object()), key
-                )
-            else:
-                return None
-
-        try:
-            # to cope with all types
-            field = cast(DictionaryObject, field.indirect_reference.get_object())  # type: ignore
-        except Exception as exc:
-            raise ValueError("field type is invalid") from exc
-        if _get_inherited(field, "/FT") is None:
-            raise ValueError("field is not valid")
-        ret = []
-        if field.get("/Subtype", "") == "/Widget":
-            if "/P" in field:
-                ret = [field["/P"].get_object()]
-            else:
-                ret = [
-                    p
-                    for p in self.pages
-                    if field.indirect_reference in p.get("/Annots", "")
-                ]
-        else:
-            kids = field.get("/Kids", ())
-            for k in kids:
-                k = k.get_object()
-                if (k.get("/Subtype", "") == "/Widget") and ("/T" not in k):
-                    # Kid that is just a widget, not a field:
-                    if "/P" in k:
-                        ret += [k["/P"].get_object()]
-                    else:
-                        ret += [
-                            p
-                            for p in self.pages
-                            if k.indirect_reference in p.get("/Annots", "")
-                        ]
-        return [
-            x
-            if isinstance(x, PageObject)
-            else (self.pages[self._get_page_number_by_indirect(x.indirect_reference)])  # type: ignore
-            for x in ret
-        ]
-
     def clone_reader_document_root(self, reader: PdfReader) -> None:
         """
         Copy the reader document root to the writer and all sub elements,
@@ -1100,61 +1030,13 @@ class PdfWriter:
         self._root = self._root_object.indirect_reference  # type: ignore[assignment]
         self._pages = self._root_object.raw_get("/Pages")
         self._flatten()
+        assert self.flattened_pages is not None
         for p in self.flattened_pages:
-            o = p.get_object()
-            self._objects[p.idnum - 1] = PageObject(self, p)
-            self._objects[p.idnum - 1].update(o.items())
-        self._root_object[NameObject("/Pages")][  # type: ignore[index]
+            p[NameObject("/Parent")] = self._pages
+            self._objects[cast(IndirectObject, p.indirect_reference).idnum - 1] = p
+        cast(DictionaryObject, self._pages.get_object())[
             NameObject("/Kids")
-        ] = self.flattened_pages
-        del self.flattened_pages
-
-    def _flatten(
-        self,
-        pages: Union[None, DictionaryObject, PageObject] = None,
-        inherit: Optional[Dict[str, Any]] = None,
-        indirect_reference: Optional[IndirectObject] = None,
-    ) -> None:
-        inheritable_page_attributes = (
-            NameObject(PG.RESOURCES),
-            NameObject(PG.MEDIABOX),
-            NameObject(PG.CROPBOX),
-            NameObject(PG.ROTATE),
-        )
-        if inherit is None:
-            inherit = {}
-        if pages is None:
-            pages = cast(DictionaryObject, self._root_object["/Pages"])
-            self.flattened_pages = ArrayObject()
-        assert pages is not None  # hint for mypy
-
-        if PA.TYPE in pages:
-            t = str(pages[PA.TYPE])
-        # if pdf has no type, considered as a page if /Kids is missing
-        elif PA.KIDS not in pages:
-            t = "/Page"
-        else:
-            t = "/Pages"
-
-        if t == "/Pages":
-            for attr in inheritable_page_attributes:
-                if attr in pages:
-                    inherit[attr] = pages[attr]
-            for page in cast(ArrayObject, pages[PA.KIDS]):
-                addt = {}
-                if isinstance(page, IndirectObject):
-                    addt["indirect_reference"] = page
-                self._flatten(page.get_object(), inherit, **addt)
-        elif t == "/Page":
-            for attr_in, value in list(inherit.items()):
-                # if the page has it's own value, it does not inherit the
-                # parent's value:
-                if attr_in not in pages:
-                    pages[attr_in] = value
-            pages[NameObject("/Parent")] = cast(
-                IndirectObject, self._root_object.raw_get("/Pages")
-            )
-            self.flattened_pages.append(indirect_reference)
+        ] = ArrayObject([p.indirect_reference for p in self.flattened_pages])
 
     def clone_document_from_reader(
         self,
@@ -1176,8 +1058,9 @@ class PdfWriter:
                 document.
         """
         self.clone_reader_document_root(reader)
+        self._info_obj = self._add_object(DictionaryObject())
         if TK.INFO in reader.trailer:
-            self._info = reader._info.clone(self).indirect_reference  # type: ignore
+            self._info = reader._info  # actually copy fields
         try:
             self._ID = cast(ArrayObject, reader._ID).clone(self)
         except AttributeError:
@@ -1321,7 +1204,7 @@ class PdfWriter:
 
     def _write_pdf_structure(self, stream: StreamType) -> List[int]:
         object_positions = []
-        stream.write(self.pdf_header + b"\n")
+        stream.write(self.pdf_header.encode() + b"\n")
         stream.write(b"%\xE2\xE3\xCF\xD3\n")
 
         for i, obj in enumerate(self._objects):
@@ -1358,7 +1241,7 @@ class PdfWriter:
             {
                 NameObject(TK.SIZE): NumberObject(len(self._objects) + 1),
                 NameObject(TK.ROOT): self._root,
-                NameObject(TK.INFO): self._info,
+                NameObject(TK.INFO): self._info_obj,
             }
         )
         if self._ID:
@@ -1383,7 +1266,8 @@ class PdfWriter:
             if isinstance(value, PdfObject):
                 value = value.get_object()
             args[NameObject(key)] = create_string_object(str(value))
-        cast(DictionaryObject, self._info.get_object()).update(args)
+        assert isinstance(self._info, DictionaryObject)
+        self._info.update(args)
 
     def _sweep_indirect_references(
         self,
@@ -1590,41 +1474,6 @@ class PdfWriter:
         Each element is a dictionaries with ``/F`` and ``/I`` keys.
         """
         return self.get_threads_root()
-
-    def get_named_dest_root(self) -> ArrayObject:
-        if CA.NAMES in self._root_object and isinstance(
-            self._root_object[CA.NAMES], DictionaryObject
-        ):
-            names = cast(DictionaryObject, self._root_object[CA.NAMES])
-            names_ref = names.indirect_reference
-            if CA.DESTS in names and isinstance(names[CA.DESTS], DictionaryObject):
-                # 3.6.3 Name Dictionary (PDF spec 1.7)
-                dests = cast(DictionaryObject, names[CA.DESTS])
-                dests_ref = dests.indirect_reference
-                if CA.NAMES in dests:
-                    # TABLE 3.33 Entries in a name tree node dictionary
-                    nd = cast(ArrayObject, dests[CA.NAMES])
-                else:
-                    nd = ArrayObject()
-                    dests[NameObject(CA.NAMES)] = nd
-            else:
-                dests = DictionaryObject()
-                dests_ref = self._add_object(dests)
-                names[NameObject(CA.DESTS)] = dests_ref
-                nd = ArrayObject()
-                dests[NameObject(CA.NAMES)] = nd
-
-        else:
-            names = DictionaryObject()
-            names_ref = self._add_object(names)
-            self._root_object[NameObject(CA.NAMES)] = names_ref
-            dests = DictionaryObject()
-            dests_ref = self._add_object(dests)
-            names[NameObject(CA.DESTS)] = dests_ref
-            nd = ArrayObject()
-            dests[NameObject(CA.NAMES)] = nd
-
-        return nd
 
     def add_outline_item_destination(
         self,
