+++ tests/scripts/test_make_release.py
@@ -20,9 +20,9 @@ COMMITS__VERSION_4_0_1 = DATA_PATH.joinpath("commits__version_4_0_1.json")
 def test_get_git_commits_since_tag():
     make_release = pytest.importorskip("make_release")
 
-    with open(COMMITS__VERSION_4_0_1, mode="rb") as commits, \
-            mock.patch("urllib.request.urlopen", side_effect=lambda n: commits), \
-            mock.patch("subprocess.check_output", return_value=GIT_LOG__VERSION_4_0_1):
+    with open(COMMITS__VERSION_4_0_1, mode="rb") as commits, mock.patch(
+        "urllib.request.urlopen", side_effect=lambda n: commits
+    ), mock.patch("subprocess.check_output", return_value=GIT_LOG__VERSION_4_0_1):
         commits = make_release.get_git_commits_since_tag("4.0.1")
     assert commits == [
         make_release.Change(
@@ -30,35 +30,35 @@ def test_get_git_commits_since_tag():
             prefix="BUG",
             message=" Use NumberObject for /Border elements of annotations (#2451)",
             author="rsinger417",
-            author_login="rsinger417"
+            author_login="rsinger417",
         ),
         make_release.Change(
             commit_hash="8cacb0fc8fee9920b0515d1289e6ee8191eb3f21",
             prefix="DOC",
             message=" Document easier way to update metadata (#2454)",
             author="Stefan",
-            author_login="stefan6419846"
+            author_login="stefan6419846",
         ),
         make_release.Change(
             commit_hash="3fb63f7e3839ce39ac98978c996f3086ba230a20",
             prefix="TST",
             message=" Avoid catching not emitted warnings (#2429)",
             author="Stefan",
-            author_login="stefan6419846"
+            author_login="stefan6419846",
         ),
         make_release.Change(
             commit_hash="61b73d49778e8f0fb172d5323e67677c9974e420",
             prefix="DOC",
             message=" Typo `Polyline` → `PolyLine` in adding-pdf-annotations.md (#2426)",
             author="CWKSC",
-            author_login="CWKSC"
+            author_login="CWKSC",
         ),
         make_release.Change(
             commit_hash="f851a532a5ec23b572d86bd7185b327a3fac6b58",
             prefix="DEV",
             message=" Bump codecov/codecov-action from 3 to 4 (#2430)",
             author="dependabot[bot]",
-            author_login="dependabot[bot]"
+            author_login="dependabot[bot]",
         ),
     ]
 
@@ -66,12 +66,14 @@ def test_get_git_commits_since_tag():
 def test_get_formatted_changes():
     make_release = pytest.importorskip("make_release")
 
-    with open(COMMITS__VERSION_4_0_1, mode="rb") as commits, \
-            mock.patch("urllib.request.urlopen", side_effect=lambda n: commits), \
-            mock.patch("subprocess.check_output", return_value=GIT_LOG__VERSION_4_0_1):
+    with open(COMMITS__VERSION_4_0_1, mode="rb") as commits, mock.patch(
+        "urllib.request.urlopen", side_effect=lambda n: commits
+    ), mock.patch("subprocess.check_output", return_value=GIT_LOG__VERSION_4_0_1):
         output, output_with_user = make_release.get_formatted_changes("4.0.1")
 
-    assert output == """
+    assert (
+        output
+        == """
 ### Bug Fixes (BUG)
 -  Use NumberObject for /Border elements of annotations (#2451)
 
@@ -85,7 +87,10 @@ def test_get_formatted_changes():
 ### Testing (TST)
 -  Avoid catching not emitted warnings (#2429)
 """
-    assert output_with_user == """
+    )
+    assert (
+        output_with_user
+        == """
 ### Bug Fixes (BUG)
 -  Use NumberObject for /Border elements of annotations (#2451) by @rsinger417
 
@@ -99,3 +104,4 @@ def test_get_formatted_changes():
 ### Testing (TST)
 -  Avoid catching not emitted warnings (#2429) by @stefan6419846
 """
+    )
