+++ tests/test_reader.py
@@ -137,7 +137,7 @@ def test_broken_meta_data(pdf_path):
         with pytest.raises(
             PdfReadError,
             match=(
-                "trailer not found or does not point to document "
+                "Trailer not found or does not point to document "
                 "information directory"
             ),
         ):
@@ -444,6 +444,7 @@ def test_get_form(src, expected, expected_get_fields, txt_file_path):
 def test_get_page_number(src, page_number):
     src = RESOURCE_ROOT / src
     reader = PdfReader(src)
+    reader._get_page(0)
     page = reader.pages[page_number]
     assert reader.get_page_number(page) == page_number
